End basics instructions
I added all the basics instructions (add, sub, mul, div, and, or, xor, shl, shr). 
I also added the instructions to tests de registers or the numbers (slt, sle and seq).
I have not tested all the instructions yet.
I also created a fonction set_regs. This functions is used to set all the value of the registers according to them number (R0 = 0, R1 = 1, ..., R31 = 31);
I created a fonction to view the registers. Now we just need to call it to show all the registers. 
To finish I created a fonction run. This function is used to run the program. In fact, in the main we juste have this fonction run. 
In this function, we first show all the values. 
After that while running is at 1, we carry out the instructions. 
The only one solution to stop the code is to put running at 0. We do it when we encounter the instruction 0x0000000. After that we put runnign at 0 (in the case 0). And we finish the program by showing alla the registers.